+++ packages/next-intl/.size-limit.ts
@@ -60,7 +60,7 @@ const config: SizeLimitConfig = [
   {
     name: 'import createMiddleware from \'next-intl/middleware\'',
     path: 'dist/production/middleware.js',
-    limit: '9.595 KB'
+    limit: '9.61 KB'
   },
   {
     name: 'import * from \'next-intl/routing\'',